chore: split relationships tests into smaller units
